style: tiptap component
